move definition to types
