+++ lib/platform/bitbucket/index.spec.ts
@@ -2,7 +2,7 @@ import URL from 'url';
 import responses from './__fixtures__/responses';
 import { GotApi, RepoParams, Platform } from '../common';
 import { REPOSITORY_DISABLED } from '../../constants/error-messages';
-import { BranchStatus } from '../../constants/branch-constants';
+import { BranchStatus } from '../../types/branch-status';
 
 describe('platform/bitbucket', () => {
   let bitbucket: Platform;
