+++ lib/platform/gitea/index.spec.ts
@@ -17,7 +17,7 @@ import {
   Platform,
 } from '..';
 import { logger as _logger } from '../../logger';
-import { BranchStatus } from '../../constants/branch-constants';
+import { BranchStatus } from '../../types/branch-status';
 import { GiteaGotApi } from './gitea-got-wrapper';
 import { CommitFilesConfig, File } from '../git/storage';
