+++ lib/types/index.ts
@@ -1,2 +1,3 @@
 export * from './host-rules';
 export * from './versioning';
+export * from './branch-status';
